d3 airport map emphasize mouse-over
I hi-light the airport when the mouse pointer hovers by outlining in
white, increasing the circle outline, and increasing opacity.  I
included a legend for the number of flights.  I made the title more
clearly emphasize the relationship between number of flights and delays.

Maximum flight delay is shown in hours.  Minutes are formatted to 1
decimal, and large numbers are formatted to integers with commas.